yo
git-svn-id: https://mason.googlecode.com/svn/trunk@519 e22d1478-1c87-2d72-604f-1c9cf2634165

+++ mason/sim/display3d/SelectionBehavior.java
@@ -45,7 +45,7 @@ public class SelectionBehavior extends MouseBehavior
     boolean oneClick = true;
     boolean twoClicks = true;
     
-    public void setTolerance(float tolerance) { pickCanvas.setTolerance(tolerance); }
+    public void setTolerance(double tolerance) { pickCanvas.setTolerance((float)tolerance); }
     
 	/** Detaches the Selection Behavior from the Canvas properly, so leaks don't occur. */
     public void detach()
