+++ mason/sim/portrayal3d/simple/CubePortrayal3D.java
@@ -19,7 +19,7 @@ import java.awt.*;
  */
 public class CubePortrayal3D extends SimplePortrayal3D
     {
-	float scale = 1f;
+	double scale = 1f;
 	Appearance appearance;
 	boolean generateNormals;
 	boolean generateTextureCoordinates;
@@ -31,7 +31,7 @@ public class CubePortrayal3D extends SimplePortrayal3D
         }
         
     /** Constructs a CubePortrayal3D with a default (flat opaque white) appearance and the given scale. */
-    public CubePortrayal3D(float scale)
+    public CubePortrayal3D(double scale)
         {
         this(Color.white,scale);
         }
@@ -43,7 +43,7 @@ public class CubePortrayal3D extends SimplePortrayal3D
         }
         
     /** Constructs a CubePortrayal3D with a flat opaque appearance of the given color and the given scale. */
-    public CubePortrayal3D(Color color, float scale)
+    public CubePortrayal3D(Color color, double scale)
         {
         this(appearanceForColor(color), false, false, scale);
         }
@@ -55,20 +55,20 @@ public class CubePortrayal3D extends SimplePortrayal3D
         }
 
     /** Constructs a CubePortrayal3D with the given (opaque) image and scale. */
-    public CubePortrayal3D(Image image, float scale)
+    public CubePortrayal3D(Image image, double scale)
         {
         this(appearanceForImage(image,true),false,true,scale);
         }
 
     /** Constructs a CubePortrayal3D with the given appearance and scale, plus whether or not to generate normals or texture coordinates.  Without texture coordiantes, a texture will not be displayed */
-    public CubePortrayal3D(Appearance appearance, boolean generateNormals, boolean generateTextureCoordinates, float scale)
+    public CubePortrayal3D(Appearance appearance, boolean generateNormals, boolean generateTextureCoordinates, double scale)
         {
         this.generateNormals = generateNormals;
         this.generateTextureCoordinates = generateTextureCoordinates;
         this.appearance = appearance;  
         this.scale = scale;
         for(int i=0;i<scaledVerts.length;i++)
-            scaledVerts[i] = verts[i]*scale;
+            scaledVerts[i] = verts[i]*(float)scale;
         }
 
     final float[] scaledVerts = new float[verts.length];
