+++ mason/sim/portrayal3d/simple/CylinderPortrayal3D.java
@@ -26,7 +26,7 @@ public class CylinderPortrayal3D extends PrimitivePortrayal3D
         }
         
     /** Constructs a CylinderPortrayal3D with a default (flat opaque white) appearance and the given scale. */
-    public CylinderPortrayal3D(float scale)
+    public CylinderPortrayal3D(double scale)
         {
         this(Color.white,scale);
         }
@@ -38,7 +38,7 @@ public class CylinderPortrayal3D extends PrimitivePortrayal3D
         }
         
     /** Constructs a CylinderPortrayal3D with a flat opaque appearance of the given color and the given scale. */
-    public CylinderPortrayal3D(Color color, float scale)
+    public CylinderPortrayal3D(Color color, double scale)
         {
         this(appearanceForColor(color),true,false,scale);
         }
@@ -50,14 +50,14 @@ public class CylinderPortrayal3D extends PrimitivePortrayal3D
         }
 
     /** Constructs a CylinderPortrayal3D with the given (opaque) image and scale. */
-    public CylinderPortrayal3D(Image image, float scale)
+    public CylinderPortrayal3D(Image image, double scale)
         {
         this(appearanceForImage(image,true),false,true,scale);
         }
 
 
     /** Constructs a CylinderPortrayal3D with the given appearance and scale, plus whether or not to generate normals or texture coordinates.  Without texture coordiantes, a texture will not be displayed. */
-    public CylinderPortrayal3D(Appearance appearance, boolean generateNormals, boolean generateTextureCoordinates, float scale)
+    public CylinderPortrayal3D(Appearance appearance, boolean generateNormals, boolean generateTextureCoordinates, double scale)
         {
         this.appearance = appearance;
         setScale(null, scale);
